requirements.txt
streamlit
google-generativeai
pandas
openpyxl
Pillow
PyMuPDF
python-dotenv
reportlab